LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/events/015-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:49001/echo_raw' failed: Unrecognized frame opcode: 3

PASS WebSockets: instanceof on events 
